+++ denops/@denops-private/service/deps.ts
@@ -1,0 +1,16 @@
+export * as flags from "https://deno.land/std@0.93.0/flags/mod.ts";
+
+export { Session } from "https://deno.land/x/msgpack_rpc@v3.1.0/mod.ts";
+
+export type { Message as VimMessage } from "https://deno.land/x/vim_channel_command@v0.6.0/mod.ts";
+export { Session as VimSession } from "https://deno.land/x/vim_channel_command@v0.6.0/mod.ts";
+
+export {
+  WorkerReader,
+  WorkerWriter,
+} from "https://deno.land/x/workerio@v1.3.1/mod.ts";
+
+export * from "https://deno.land/x/unknownutil@v0.1.1/mod.ts";
+
+export { using } from "https://deno.land/x/disposable@v0.2.0/mod.ts";
+export type { Disposable } from "https://deno.land/x/disposable@v0.2.0/mod.ts";
